+++ web/src/main/java/com/bestomb/common/response/music/MusicVo.java
@@ -7,7 +7,7 @@ public class MusicVo {
 
     private String fileAddress;
 
-    private Long fileSize;
+    private Integer fileSize;
 
     public String getId() {
         return id;
@@ -33,12 +33,11 @@ public class MusicVo {
         this.fileAddress = fileAddress == null ? null : fileAddress.trim();
     }
 
-    public Long getFileSize() {
+    public Integer getFileSize() {
         return fileSize;
     }
 
-    public void setFileSize(Long fileSize) {
+    public void setFileSize(Integer fileSize) {
         this.fileSize = fileSize;
     }
-
 }
